Add runtime analysis and display of results for detection and tracking, add break condition for directory frame sources, fix colors when displaying
A break condition for frame sources of type DirectoryFrameSource was added. Thus, the tracking is stopped as soon as all frames have been processed. Condition checks whether the total number of processed frames equals the number of frames in the source. For frame sources of type ServerFrameSource, the number of frames is set to inifinity. Thereby, the script can only be stopped via Ctlr + C.